requirements.txt
PyQt6~=6.7.1
pyqtgraph~=0.13.7
platformdirs~=4.2.2
python-a2s==1.3.0